time_tracker-v13/requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
psycopg2-binary
python-multipart
python-jose
passlib[bcrypt]
openpyxl
openai
pytest
celery
redis
python-docx
pdfplumber